requirements.txt
gspread==6.1.2
supabase==2.7.4
requests==2.32.3
python-dotenv==1.0.1

